pyproject.toml
[project]
name = "chief"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Batista Harahap", email = "batista@bango29.com" }
]
dependencies = [
    "pydantic-ai[logfire]>=1.0.1",
    "pydantic-settings>=2.10.1",
    "typer>=0.17.4",
    "logfire>=4.4.0",
    "pydantic-graph>=1.0.1",
    "beanie>=2.0.0",
    "coolname>=2.2.0",
    "tavily-python>=0.7.11",
    "asyncer>=0.0.8",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.12.12",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-env>=1.1.5",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [
    "src/chief",
    "src/libagentic",
    "src/libshared",
]

[project.scripts]
chief = "appclis.chief:app"
chen = "appclis.chen:app"
